+++ Kernel/include/memoryDriver.h
@@ -2,7 +2,7 @@
 #define MEM_DRIVER
 #include "../include/lib.h"
 
-void * mallocFun(unsigned nbytes);
+void * mallocFun(uint64_t nbytes);
 void freeFun(void * block);
 void consult(char* buf);
