+++ Kernel/include/memoryDriverPropio.h
@@ -1,6 +1,8 @@
+#include "../include/lib.h"
 #ifndef MEM_DRIVER_PROPIO
 #define MEM_DRIVER_PROPIO
-void * mallocFun(unsigned nbytes);
+
+void * mallocFun(uint64_t nbytes);
 void freeFun(void * block);
 void consult(char* buf);
